move MapFeedbackChip,SharePositionChip to module
move MapFeedbackChip to module/feedback
move SharePositionChip tp module/share
