+++ src/modules/feedback/components/assistChip/MapFeedbackChip.js
@@ -1,12 +1,12 @@
 /**
  * @module modules/map/components/assistChips/MapFeedbackChip
  */
-import { html } from '../../../../../node_modules/lit-html/lit-html';
+import { html } from 'lit-html';
 import { $injector } from '../../../../injection/index';
 import { closeModal, openModal } from '../../../../store/modal/modal.action';
 import { isCoordinate } from '../../../../utils/checks';
 import { AbstractAssistChip } from '../../../chips/components/assistChips/AbstractAssistChip';
-import { FeedbackType } from '../../../feedback/components/toggleFeedback/ToggleFeedbackPanel';
+import { FeedbackType } from '../toggleFeedback/ToggleFeedbackPanel';
 import mapFeedbackIcon from './assets/map.svg';
 
 const Update = 'update';
