+++ src/modules/feedback/index.js
@@ -3,6 +3,7 @@ import { MapFeedbackPanel } from './components/mapFeedback/MapFeedbackPanel';
 import { ToggleFeedbackPanel } from './components/toggleFeedback/ToggleFeedbackPanel';
 import { GeneralFeedbackPanel } from './components/generalFeedback/GeneralFeedbackPanel';
 import { LikertItemRatingPanel } from './components/rating/LikertItemRatingPanel';
+import { MapFeedbackChip } from './components/assistChip/MapFeedbackChip';
 
 if (!window.customElements.get(MapFeedbackPanel.tag)) {
 	window.customElements.define(MapFeedbackPanel.tag, MapFeedbackPanel);
@@ -18,4 +19,8 @@ if (!window.customElements.get(GeneralFeedbackPanel.tag)) {
 
 if (!window.customElements.get(LikertItemRatingPanel.tag)) {
 	window.customElements.define(LikertItemRatingPanel.tag, LikertItemRatingPanel);
+}
+
+if (!window.customElements.get(MapFeedbackChip.tag)) {
+	window.customElements.define(MapFeedbackChip.tag, MapFeedbackChip);
 }
