+++ src/modules/map/index.js
@@ -5,4 +5,3 @@ import './components/geolocationButton';
 import './components/contextMenu';
 import './components/attributionInfo';
 import './components/rotationButton';
-import './components/assistChips';
