+++ src/modules/share/components/assistChip/ShareDataChip.js
@@ -3,7 +3,7 @@
  */
 import { html } from 'lit-html';
 import { QueryParameters } from '../../../../domain/queryParameters';
-import { $injector } from '../../../../injection';
+import { $injector } from '../../../../injection/index';
 import { openModal } from '../../../../store/modal/modal.action';
 import { AbstractAssistChip } from '../../../chips/components/assistChips/AbstractAssistChip';
 import shareIcon from './assets/share.svg';
