+++ test/modules/feedback/components/assistChip/MapFeedbackChip.test.js
@@ -1,7 +1,7 @@
 import { $injector } from '../../../../../src/injection';
 import { FeedbackType, ToggleFeedbackPanel } from '../../../../../src/modules/feedback/components/toggleFeedback/ToggleFeedbackPanel';
-import { MapFeedbackChip } from '../../../../../src/modules/map/components/assistChips/MapFeedbackChip';
-import mapSvg from '../../../../../src/modules/map/components/assistChips/assets/map.svg';
+import { MapFeedbackChip } from '../../../../../src/modules/feedback/components/assistChip/MapFeedbackChip';
+import mapSvg from '../../../../../src/modules/feedback/components/assistChip/assets/map.svg';
 import { closeModal } from '../../../../../src/store/modal/modal.action';
 import { modalReducer } from '../../../../../src/store/modal/modal.reducer';
 import { TestUtils } from '../../../../test-utils';
