+++ test/modules/share/components/assistChip/SharePositionChip.test.js
@@ -1,7 +1,7 @@
 import { QueryParameters } from '../../../../../src/domain/queryParameters';
 import { $injector } from '../../../../../src/injection';
-import { SharePositionChip } from '../../../../../src/modules/map/components/assistChips/SharePositionChip';
-import shareSvg from '../../../../../src/modules/map/components/assistChips/assets/share.svg';
+import { SharePositionChip } from '../../../../../src/modules/share/components/assistChip/SharePositionChip';
+import shareSvg from '../../../../../src/modules/share/components/assistChip/assets/share.svg';
 import { ShareDialogContent } from '../../../../../src/modules/share/components/dialog/ShareDialogContent';
 import { modalReducer } from '../../../../../src/store/modal/modal.reducer';
 import { LevelTypes } from '../../../../../src/store/notifications/notifications.action';
